Refactor: Move ingest logic to dedicated modules
Extract helper functions into new modules for better organization and testability.
