+++ tests/unit/ingest/test_utils.py
@@ -1,16 +1,20 @@
 import tempfile
 import os
-from ai_proxy.logdb.ingest import (
-    _safe_iso_to_datetime,
-    _file_sha256,
-    _file_prefix_sha256,
-    _iter_json_blocks,
-    _parse_log_entry,
-    _normalize_entry,
-    _estimate_batch_bytes,
-    _env_int,
-    _scan_log_file,
-    _derive_server_id
+from ai_proxy.logdb.parsers.log_parser import (
+    iter_json_blocks,
+    parse_log_entry,
+    normalize_entry,
+)
+from ai_proxy.logdb.utils.datetime_utils import safe_iso_to_datetime
+from ai_proxy.logdb.utils.file_utils import (
+    file_sha256,
+    file_prefix_sha256,
+    env_int,
+)
+from ai_proxy.logdb.utils.server_utils import derive_server_id
+from ai_proxy.logdb.processing.batch_processor import (
+    estimate_batch_bytes,
+    scan_log_file,
 )
 from tests.unit.shared.ingest_fixtures import SAMPLE_ENTRY_1
 
@@ -19,12 +23,12 @@ def test_safe_iso_to_datetime_edge_cases():
     """Test _safe_iso_to_datetime with various edge cases."""
 
     # Test empty string (line 26)
-    assert _safe_iso_to_datetime("") is None
-    assert _safe_iso_to_datetime(None) is None
+    assert safe_iso_to_datetime("") is None
+    assert safe_iso_to_datetime(None) is None
 
     # Test invalid format
-    assert _safe_iso_to_datetime("invalid-date") is None
-    assert _safe_iso_to_datetime("2025-13-45T25:70:80Z") is None
+    assert safe_iso_to_datetime("invalid-date") is None
+    assert safe_iso_to_datetime("2025-13-45T25:70:80Z") is None
 
 
 def test_file_sha256_function(tmp_path):
@@ -35,7 +39,7 @@ def test_file_sha256_function(tmp_path):
     test_file.write_bytes(b"Hello, World!")
 
     # Test SHA256 calculation
-    sha = _file_sha256(str(test_file))
+    sha = file_sha256(str(test_file))
     assert len(sha) == 64  # SHA256 hex digest length
     assert sha == "dffd6021bb2bd5b0af676290809ec3a53191dd81c7f70a4b28688a362182986f"
 
@@ -48,11 +52,11 @@ def test_file_prefix_sha256_edge_cases(tmp_path):
     test_file.write_bytes(b"Hi")
 
     # Test reading more bytes than available (should break early)
-    sha = _file_prefix_sha256(str(test_file), 1000000)  # Much larger than file
+    sha = file_prefix_sha256(str(test_file), 1000000)  # Much larger than file
     assert len(sha) == 64
 
     # Test zero bytes
-    sha_zero = _file_prefix_sha256(str(test_file), 0)
+    sha_zero = file_prefix_sha256(str(test_file), 0)
     assert len(sha_zero) == 64
 
 
@@ -64,7 +68,7 @@ def test_iter_json_blocks_no_braces(tmp_path):
     test_file.write_text("2025-09-10 12:00:00 - INFO - No JSON here\nAnother line without braces\n")
 
     with open(test_file, "r") as f:
-        blocks = list(_iter_json_blocks(f))
+        blocks = list(iter_json_blocks(f))
         assert len(blocks) == 0  # Should find no JSON blocks
 
 
@@ -72,19 +76,19 @@ def test_parse_log_entry_edge_cases():
     """Test _parse_log_entry with various invalid inputs (lines 160-161, 165, 167)."""
 
     # Test invalid JSON (line 160-161)
-    assert _parse_log_entry("{ invalid json") is None
-    assert _parse_log_entry("") is None
+    assert parse_log_entry("{ invalid json") is None
+    assert parse_log_entry("") is None
 
     # Test non-dict JSON (line 165)
-    assert _parse_log_entry("[]") is None
-    assert _parse_log_entry("\"string\"") is None
-    assert _parse_log_entry("123") is None
+    assert parse_log_entry("[]") is None
+    assert parse_log_entry("\"string\"") is None
+    assert parse_log_entry("123") is None
 
     # Test missing required fields (line 167)
-    assert _parse_log_entry("{}") is None
-    assert _parse_log_entry('{"endpoint": "/v1/chat"}') is None
-    assert _parse_log_entry('{"request": {}}') is None
-    assert _parse_log_entry('{"response": {}}') is None
+    assert parse_log_entry("{}") is None
+    assert parse_log_entry('{"endpoint": "/v1/chat"}') is None
+    assert parse_log_entry('{"request": {}}') is None
+    assert parse_log_entry('{"response": {}}') is None
 
 
 def test_normalize_entry_edge_cases():
@@ -92,22 +96,22 @@ def test_normalize_entry_edge_cases():
 
     # Test missing timestamp (line 175)
     entry_no_ts = {"endpoint": "/v1/chat", "request": {}, "response": {}}
-    assert _normalize_entry(entry_no_ts) is None
+    assert normalize_entry(entry_no_ts) is None
 
     # Test invalid timestamp (line 182)
     entry_bad_ts = {"timestamp": "invalid", "endpoint": "/v1/chat", "request": {}, "response": {}}
-    assert _normalize_entry(entry_bad_ts) is None
+    assert normalize_entry(entry_bad_ts) is None
 
     # Test empty endpoint (line 187)
     entry_no_endpoint = {"timestamp": "2025-09-10T12:00:00Z", "endpoint": "", "request": {}, "response": {}}
-    assert _normalize_entry(entry_no_endpoint) is None
+    assert normalize_entry(entry_no_endpoint) is None
 
     # Test missing request/response (line 192)
     entry_no_req = {"timestamp": "2025-09-10T12:00:00Z", "endpoint": "/v1/chat", "response": {}}
-    assert _normalize_entry(entry_no_req) is None
+    assert normalize_entry(entry_no_req) is None
 
     entry_no_resp = {"timestamp": "2025-09-10T12:00:00Z", "endpoint": "/v1/chat", "request": {}}
-    assert _normalize_entry(entry_no_resp) is None
+    assert normalize_entry(entry_no_resp) is None
 
     # Test unserializable request/response (line 197-198)
     class UnserializableObj:
@@ -120,7 +124,7 @@ def test_normalize_entry_edge_cases():
         "request": UnserializableObj(),
         "response": {}
     }
-    assert _normalize_entry(entry_bad_json) is None
+    assert normalize_entry(entry_bad_json) is None
 
 
 def test_normalize_entry_invalid_status_and_latency():
@@ -134,7 +138,7 @@ def test_normalize_entry_invalid_status_and_latency():
         "response": {},
         "status_code": "not-a-number"
     }
-    result = _normalize_entry(entry_bad_status)
+    result = normalize_entry(entry_bad_status)
     assert result is not None
     assert result["status_code"] is None
 
@@ -146,7 +150,7 @@ def test_normalize_entry_invalid_status_and_latency():
         "response": {},
         "latency_ms": "not-a-float"
     }
-    result = _normalize_entry(entry_bad_latency)
+    result = normalize_entry(entry_bad_latency)
     assert result is not None
     assert result["latency_ms"] is None
 
@@ -162,7 +166,7 @@ def test_estimate_batch_bytes_exception_handling():
     ]
 
     # Should handle exceptions gracefully
-    total_bytes = _estimate_batch_bytes(problematic_batch)
+    total_bytes = estimate_batch_bytes(problematic_batch)
     assert total_bytes > 0  # Should include overhead even if some rows fail
 
 
@@ -171,11 +175,11 @@ def test_env_int_exception_handling(monkeypatch):
 
     # Test with invalid environment variable
     monkeypatch.setenv("TEST_INVALID_INT", "not-a-number")
-    result = _env_int("TEST_INVALID_INT", 42)
+    result = env_int("TEST_INVALID_INT", 42)
     assert result == 42  # Should return default
 
     # Test with missing environment variable
-    result = _env_int("TEST_MISSING_VAR", 99)
+    result = env_int("TEST_MISSING_VAR", 99)
     assert result == 99  # Should return default
 
 
@@ -203,7 +207,7 @@ def test_scan_log_file_sha_validation_exception_coverage():
 
 def test_scan_log_file_seek_logic(tmp_path):
     """Test file seek logic in resume (lines 350-353)."""
-    from ai_proxy.logdb.ingest import _derive_server_id
+    # derive_server_id is now imported at the top
 
     logs_dir = tmp_path / "logs"
     db_base = tmp_path / "logs" / "db"
@@ -214,33 +218,33 @@ def test_scan_log_file_seek_logic(tmp_path):
     content = "First line\n" + SAMPLE_ENTRY_1 + "Last line\n"
     log_path.write_text(content, encoding="utf-8")
 
-    server_id = _derive_server_id(str(db_base))
+    server_id = derive_server_id(str(db_base))
 
     # This will exercise the seek logic when resuming
-    inserted, skipped = _scan_log_file(str(log_path), str(db_base), None, None, server_id)
+    inserted, skipped = scan_log_file(str(log_path), str(db_base), None, None, server_id)
     assert inserted >= 0  # Should complete without error
 
 
 def test_derive_server_id_edge_cases(tmp_path, monkeypatch):
-    """Test _derive_server_id with various scenarios."""
-    from ai_proxy.logdb.ingest import _derive_server_id
+    """Test derive_server_id with various scenarios."""
+    # derive_server_id is now imported at the top
     import os
 
     # Test with explicit LOGDB_SERVER_ID env var
     monkeypatch.setenv("LOGDB_SERVER_ID", "explicit-server-123")
-    server_id = _derive_server_id(str(tmp_path))
+    server_id = derive_server_id(str(tmp_path))
     assert server_id == "explicit-server-123"
 
     # Test without base_db_dir
     monkeypatch.delenv("LOGDB_SERVER_ID", raising=False)
-    server_id = _derive_server_id(None)
+    server_id = derive_server_id(None)
     assert len(server_id) == 36  # UUID format
 
     # Test with unwritable directory (should fall back to hostname-based)
     unwritable_dir = tmp_path / "unwritable"
     unwritable_dir.mkdir(mode=0o000)  # No write permissions
     try:
-        server_id = _derive_server_id(str(unwritable_dir))
+        server_id = derive_server_id(str(unwritable_dir))
         assert len(server_id) == 36  # Should still generate UUID
     finally:
         unwritable_dir.chmod(0o755)  # Restore permissions for cleanup
@@ -251,11 +255,11 @@ def test_derive_server_id_edge_cases(tmp_path, monkeypatch):
     server_file = writable_dir / ".server_id"
     server_file.write_text("existing-server-456")
 
-    server_id = _derive_server_id(str(writable_dir))
+    server_id = derive_server_id(str(writable_dir))
     assert server_id == "existing-server-456"
 
     # Test with empty server_id file (should generate new one)
     server_file.write_text("")  # Empty file
-    server_id = _derive_server_id(str(writable_dir))
+    server_id = derive_server_id(str(writable_dir))
     assert len(server_id) == 36
     assert server_id != "existing-server-456"
